Re organize services
