+++ src/scripts/services/mapContests.ts
