--- src/scripts/utilities/transformPlayers.ts
@@ -1,9 +1,0 @@
-import { IDraftKingsPlayer } from '../../interfaces/draftkings/IDraftKingsPlayer';
-
-export function transformPlayers(players: IDraftKingsPlayer[]) {
-	return players.map((player) => ({
-		...player,
-		isExcluded: false,
-		isLocked: false,
-	}));
-}
